Break up parseKMLs into many files
