+++ src/cli/scrapeRoutes/parseKMLs/MAPBOX_TOKEN.ts
@@ -1,0 +1,2 @@
+export const MAPBOX_TOKEN =
+  'pk.eyJ1Ijoic3BpbmRyaWZ0IiwiYSI6ImNqaDg2bDBsdTBmZG0yd3MwZ2x4ampsdXUifQ.7E19C7BhF9Dfd1gdJiYTEg';
