+++ src/cli/scrapeRoutes/parseKMLs/TileId.ts
@@ -1,0 +1,5 @@
+export interface TileId {
+  x: number;
+  y: number;
+  z: number;
+}
