+++ src/cli/scrapeRoutes/parseKMLs/earthCircumference.ts
@@ -1,0 +1,1 @@
+export const earthCircumference = 40075016.686;
