+++ src/cli/scrapeRoutes/parseKMLs/flattenFeatureCollection.ts
@@ -1,0 +1,9 @@
+import {Feature, FeatureCollection} from '@turf/helpers';
+
+export function flattenFeatureCollection(input: Feature | FeatureCollection): Feature[] {
+  if (input.type === 'FeatureCollection') {
+    return input.features.flatMap(flattenFeatureCollection);
+  } else {
+    return [input];
+  }
+}
